requirements.txt
pypdf
python-docx
google-generativeai
streamlit
